+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Setup fragment result listener in `onCreate` instead of `onViewCreated`
 - Report last key name when removing last key fails in `History`
 - Bump AGP to v8.1.3
+- Migrate settings screen to use Compose UI
 
 ### Changes